done backend

+++ backend/controllers/goalController.js
@@ -2,6 +2,7 @@ const asyncHandler = require('express-async-handler');
 
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc getGoals
 // @route GET /api/goals
@@ -9,7 +10,7 @@ const Goal = require('../models/goalModel');
 
 const getGoals = asyncHandler(async(req, res) => {
 
-  const goal =  await Goal.find();
+  const goal =  await Goal.find({user : req.user.id});
 
   if(goal == 0){
     res.status(400)
@@ -26,7 +27,8 @@ const getGoals = asyncHandler(async(req, res) => {
 const setGoals = asyncHandler(async(req, res) => {
 
   const goal = await Goal.create({
-    text : req.body.text
+    text : req.body.text,
+    user : req.user.id
   })
   if (!goal) {
     res.status(400);
@@ -46,6 +48,17 @@ const updateGoals = asyncHandler(async(req, res) => {
   if(!goal){
     res.status(404);
     throw new Error("Goal not found")
+  }
+  const user = await User.findById(req.user.id);
+  if(!user){
+    res.status(401);
+    throw new Error("User not found")
+  }
+  // make sure the logged in user is the owner of the goal
+  if(goal.user.toString() !== req.user.id){
+    res.status(401);
+    throw new Error("Not authorized")
+
   }
  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{
   new : true,
@@ -68,8 +81,19 @@ const deleteGoals = asyncHandler( async(req, res) => {
 
     throw new Error("Goal not found")
   }
+  const user = await User.findById(req.user.id);
+  if(!user){
+    res.status(401);
+    throw new Error("User not found")
+  }
+  // make sure the logged in user is the owner of the goal
+  if(goal.user.toString() !== req.user.id){
+    res.status(401);
+    throw new Error("Not authorized")
 
+  }
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
+  
   res.status(201).json({
     message: deletedGoal
   });
